feat: 지출 목록 조회 #1 - 기간으로 조회

+++ src/main/java/com/wealdy/saemsembackend/domain/spending/controller/response/SpendingResponse.java
@@ -14,7 +14,7 @@ public class SpendingResponse {
     private Long id;
     private String categoryName;
     private LocalDateTime date;
-    private int amount;
+    private long amount;
     private String memo;
     private YnColumn excludeTotal;
